+++ src/graphql.rs
@@ -257,9 +257,68 @@ pub trait ___Type {
     }
 }
 
-pub struct __Directive {}
+#[derive(Clone, Debug)]
+pub struct __Directive {
+    pub name: String,
+    pub description: Option<String>,
+    pub locations: Vec<__DirectiveLocation>,
+    pub args: Vec<__InputValue>,
+    pub is_repeatable: bool,
+}
 
-pub struct __DirectiveLocation {}
+impl __Directive {
+    pub const TYPE: &'static str = "__Directive";
+
+    // name: String!
+    pub fn name(&self) -> &str {
+        &self.name
+    }
+
+    // description: String
+    pub fn description(&self) -> Option<&String> {
+        self.description.as_ref()
+    }
+
+    // locations: [__DirectiveLocation!]!
+    pub fn locations(&self) -> &[__DirectiveLocation] {
+        &self.locations
+    }
+
+    // args: [__InputValue!]!
+    pub fn args(&self) -> &[__InputValue] {
+        &self.args
+    }
+
+    // isRepeatable: Boolean!
+    pub fn is_repeatable(&self) -> bool {
+        self.is_repeatable
+    }
+}
+
+#[derive(Serialize, Clone, Debug)]
+#[allow(dead_code)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
+pub enum __DirectiveLocation {
+    Query,
+    Mutation,
+    Subscription,
+    Field,
+    FragmentDefinition,
+    FragmentSpread,
+    InlineFragment,
+    VariableDefinition,
+    Schema,
+    Scalar,
+    Object,
+    FieldDefinition,
+    ArgumentDefinition,
+    Interface,
+    Union,
+    Enum,
+    EnumValue,
+    InputObject,
+    InputFieldDefinition,
+}
 
 pub trait ___Field {
     // name: String!
@@ -3754,6 +3813,51 @@ impl __Schema {
     // directives: [__Directive!]!
     #[allow(dead_code)]
     pub fn directives(&self) -> Vec<__Directive> {
-        vec![]
+        vec![
+            __Directive {
+                name: "include".to_string(),
+                description: Some(
+                    "This field or fragment will be included only when the `if` argument is true."
+                        .to_string(),
+                ),
+                locations: vec![
+                    __DirectiveLocation::Field,
+                    __DirectiveLocation::FragmentSpread,
+                    __DirectiveLocation::InlineFragment,
+                ],
+                args: vec![__InputValue {
+                    name_: "if".to_string(),
+                    type_: __Type::NonNull(NonNullType {
+                        type_: Box::new(__Type::Scalar(Scalar::Boolean)),
+                    }),
+                    description: Some("Included when true".to_string()),
+                    default_value: None,
+                    sql_type: None,
+                }],
+                is_repeatable: false,
+            },
+            __Directive {
+                name: "skip".to_string(),
+                description: Some(
+                    "This field or fragment will be skipped when the `if` argument is true."
+                        .to_string(),
+                ),
+                locations: vec![
+                    __DirectiveLocation::Field,
+                    __DirectiveLocation::FragmentSpread,
+                    __DirectiveLocation::InlineFragment,
+                ],
+                args: vec![__InputValue {
+                    name_: "if".to_string(),
+                    type_: __Type::NonNull(NonNullType {
+                        type_: Box::new(__Type::Scalar(Scalar::Boolean)),
+                    }),
+                    description: Some("Skipped when true".to_string()),
+                    default_value: None,
+                    sql_type: None,
+                }],
+                is_repeatable: false,
+            },
+        ]
     }
 }