+++ docs/changelog.md
@@ -38,3 +38,4 @@
 ## master
 - rename enum variants with comment directive `@graphql({"mappings": "sql-value": "graphql_value""})`
 - bugfix: query with more than 50 fields fails
+- bugfix: @skip and @include directives missing from introspection schema